host: Only cleanup resources that exist

+++ host/cleanup.sh
@@ -5,17 +5,21 @@
 set -e
 
 JOB_ID="$1"
+
 ZPOOL="flynn-${JOB_ID}"
+if zpool status "${ZPOOL}" &>/dev/null; then
+  # try multiple times to destroy the zpool in case it's still busy
+  for i in $(seq 10); do
+    echo "destroying zpool: ${ZPOOL}"
+    if zpool destroy "${ZPOOL}"; then
+      break
+    fi
+    sleep 1
+  done
+fi
+
 HOST_DIR="/var/lib/flynn/${JOB_ID}"
-
-# try multiple times to destroy the zpool in case it's still busy
-for i in $(seq 10); do
-  echo "destroying zpool: ${ZPOOL}"
-  if zpool destroy "${ZPOOL}"; then
-    break
-  fi
-  sleep 1
-done
-
-echo "removing host dir: ${HOST_DIR}"
-rm -rf "${HOST_DIR}"
+if [[ -d "${HOST_DIR}" ]]; then
+  echo "removing host dir: ${HOST_DIR}"
+  rm -rf "${HOST_DIR}"
+fi
